Rename set-splitting to splitting
