+++ common/src/main/java/org/valkyrienskies/mod/mixin/feature/commands/MixinArgumentTypeInfos.java
@@ -10,9 +10,9 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.valkyrienskies.mod.common.command.RelativeVector3Argument;
-import org.valkyrienskies.mod.common.command.ShipArgument;
-import org.valkyrienskies.mod.common.command.ShipArgumentInfo;
+import org.valkyrienskies.mod.common.command.arguments.RelativeVector3Argument;
+import org.valkyrienskies.mod.common.command.arguments.ShipArgument;
+import org.valkyrienskies.mod.common.command.arguments.ShipArgumentInfo;
 
 @Mixin(ArgumentTypeInfos.class)
 public class MixinArgumentTypeInfos {
